packages/research-ingestion/requirements.txt
# CompoundAtlas Research Ingestion Pipeline
httpx>=0.27.0
pydantic>=2.6.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
rich>=13.7.0        # Pretty console output for pipeline runs
typer>=0.12.0       # CLI framework
psycopg2-binary>=2.9.9  # PostgreSQL driver (for direct DB writes)
redis>=5.0.0        # Caching layer
